programs bugfixes

+++ SoulissApp/src/main/java/it/angelic/soulissclient/Constants.java
@@ -36,7 +36,7 @@ public class Constants {
 	//public static final int COMMAND_MASSIVE = 5;
     public static final short COMMAND_FAKE_SCENE = -2;
 
-	public static final int MASSIVE_NODE_ID = -1;
+	public static final short MASSIVE_NODE_ID = -1;
 	public static final int CONNECTION_NONE = -1;
 	//Souliss Data Intent
 	public static final String CUSTOM_INTENT = "it.angelic.soulissclient.GOT_DATA";
